+++ packages/adminui-pub/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.23.1] - 2022-06-19
 ## [5.23.0] - 2022-06-15
 ### Added
  - added global `resize` handler for touch devices what scroll input into viewport when on-screen keyboard appears